refactor: Event폴더 삭제 및 Event관련 항목들을 Store폴더로 통합
