+++ src/app.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common'
 import { UsersModule } from './users/users.module'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { StoresModule } from './stores/stores.module'
-import { EventsModule } from './events/events.module'
 import { ReviewsModule } from './reviews/reviews.module'
 import { typeOrmConfig } from './config/typeorm.config'
 import { MapsModule } from './maps/maps.module'
@@ -18,7 +17,6 @@ import { AuthModule } from './auth/auth.module'
     TypeOrmModule.forRoot(typeOrmConfig),
     UsersModule,
     StoresModule,
-    EventsModule,
     MapsModule,
     ReviewsModule,
     OpenaiModule,
