+++ src/stores/DTO/read-all-events.dto.ts
