+++ src/stores/DTO/read-event.dto.ts
