+++ src/stores/entities/event-status.enum.ts
