+++ src/stores/entities/event.entity.ts
@@ -1,6 +1,6 @@
 import { Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from "typeorm"
-import { EventStatus } from "./event-status.enum"
 import { Store } from "src/stores/entities/store.entity"
+import { EventStatus } from "./event-status.enum"
 
 @Entity()
 export class Event {
