+++ src/stores/entities/store.entity.ts
@@ -1,11 +1,11 @@
 import { Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, OneToMany, OneToOne, PrimaryGeneratedColumn } from "typeorm"
 import { StoreCategory } from "./store-category.enum"
 import { Review } from "src/reviews/entites/review.entity"
-import { Event } from "src/events/entities/event.entity"
 import { StoreRequest } from "../../store-requests/entities/store-request.entity"
 import { Menu } from "src/menus/entities/menu.entity"
 import { ManagerRequest } from "src/manager-requests/entities/manager-requests.entity"
 import { User } from "src/users/entities/user.entity"
+import { Event } from "./event.entity"
 
 @Entity()
 export class Store {
